Bayesian Models

UniversalDiffEq.BayesianUDE — Type

BayesianUDE

Basic data structure used to the model structure, parameters and data for Bayesian UDE and NODE models. ...

Elements

times: a vector of times for each observation

data: a matrix of observaitons at each time point

X: a DataFrame with any covariates used by the model

data_frame: a DataFrame with columns for the time of each observation and values of the state variables

parameters: a ComponentArray that stores model parameters

loss_function: the loss function used to fit the model

process_model: a Julia mutable struct used to define model predictions

process_loss: a Julia mutable struct used to measure the performance of model predictions

observation_model: a Julia mutable struct used to predict observaitons given state variable estimates

observaiton_loss: a Julia mutable struct used to measure the performance of the observation model

process_regularization: a Julia mutable struct used to store data needed for process model regularization

observation_regularization: a Julia mutable struct used to store data needed for observation model regularization

constructor: A function that initializes a UDE model with identical structure.

...

UniversalDiffEq.BayesianNODE — Function

BayesianNODE(data;kwargs ... )

Constructs a Bayesian continuous time model for the data set data using a single layer neural network to represent the systems dynamics.

BayesianNODE(data,X;kwargs ... )

When a dataframe X is supplied the model will run with covariates. the argument X should have a column for time t with the value for time in the remaining columns. The values in X will be interpolated with a linear spline for values of time not included in the data frame.

NODEs

UniversalDiffEq.BayesianNODE — Method

BayesianNODE(data;kwargs ... )

Constructs a Bayesian continuous time model for the data set data using a single layer neural network to represent the systems dynamics.

UniversalDiffEq.BayesianNODE — Method

BayesianNODE(data,X;kwargs ... )

When a dataframe X is supplied the model will run with covariates. the argument X should have a column for time t with the value for time in the remaining columns. The values in X will be interpolated with a linear spline for values of time not included in the data frame.

UDEs

UniversalDiffEq.BayesianCustomDerivatives — Method

BayesianCustomDerivatives(data,derivs!,initial_parameters;kwargs ... )

Constructs a Bayesian UDE model for the data set data based on user defined derivatives derivs. An initial guess of model parameters are supplied with the initial_parameters argument.

...

Arguments

data: a DataFrame object with the time of observations in a column labeled t and the remaining columns the value of the state variables at each time point.

derivs: a Function of the form derivs!(du,u,p,t) where u is the value of the state variables, p are the model parameters, t is time, and du is updated with the value of the derivatives

init_parameters: A NamedTuple with the model parameters. Neural network parameters must be listed under the key NN.

...

UniversalDiffEq.BayesianCustomDerivatives — Method

BayesianCustomDerivatives(data::DataFrame,derivs!::Function,initial_parameters,priors::Function;kwargs ... )

When a function's priors is supplied its value will be added to the loss function as a penalty term for user specified parameters. It should take the a single NamedTuple p as an argument penalties for each paramter should be calculated by accessing p with the period operator.

The prior function can be used to nudge the fitted model toward prior expectations for a parameter value. For example, the following function increases the loss when a parameter p.r has a value other than 1.5, nad a second parameter p.beta is greater than zeros.

function priors(p)
    l = 0.01*(p.r - 1.5)^2
    l += 0.01*(p.beta)^2
    return l
end 

UniversalDiffEq.BayesianCustomDerivatives — Method

BayesianCustomDerivatives(data::DataFrame,X,derivs!::Function,initial_parameters;kwargs ... )

When a dataframe X is supplied the model will run with covariates. the argument X should have a column for time t with the value for time in the remaining columns. The values in X will be interpolated with a linear spline for value of time not included in the data frame.

When X is provided the derivs function must have the form derivs!(du,u,x,p,t) where x is a vector with the value of the covariates at time t.