Test models

UniversalDiffEq.jl provides a number of functions to test the performance of NODE and UDE models, with both in-sample and out-of-sample data.

Evaluating model fits

There are two primary functions to test model fits: plot_state_estimates and plot_predictions. The model fitting procedure estimates the value of the state variables $\hat{u}$ at each time point in the data set as well as the parameters of the NODE or UDE model that predicts changes in the state variables. The plot_state_estimates function compares the estimated states to the data to check the quality of the state estimates, and plot_predictions compares the predictions of UDE model one step into the future to the estimated sequence of state variables. Both functions take a UDE as an input and return a plot showing the correspondence between model predictions and observations.

UniversalDiffEq.plot_state_estimates — Method

plot_state_estimates(UDE::UDE)

Plots the values of the state variables estimated by the UDE model.

UniversalDiffEq.plot_predictions — Method

plot_predictions(UDE::UDE)

Plots the correspondence between the observed state transitions and the predictions from the UDE model.

There are also functions to compare the model predictions to out-of-sample data. The simplest is plot_forecast, which compares the observations in the test data set to a deterministic simulation from the data set, which starts at the first observation and runs to the end of the test data.

UniversalDiffEq.plot_forecast — Method

plot_forecast(UDE::UDE, test_data::DataFrame)

Plots the model's forecast over the range of the test data along with the value of the test data.

It is also possible to test the performance of the models one time step into the future using the plot_predictions function. When a test data set is supplied to the plot_predictions function, it will run a series of forecasts starting at each point in the data set, predicting one time step into the future. The function returns a plot comparing the predicted and observed changes.

UniversalDiffEq.plot_predictions — Method

plot_predictions(UDE::UDE, test_data::DataFrame)

Plots the correspondence between the observed state transitions in test data and the predictions from the UDE model.

Cross-validation

Cross-validation is important for model comparison and hyper-parameter tuning. The leave_future_out_cv function breaks the data set into training and test data sets by leaving off the final observations in the data set. The model is then trained on the beginning of the data set and the performance is calculated by comparing a forecast to the test data. The user can specify the time horizon for the forecast $T_{Forecast}$ and the number of tests $K$. The first test trains the model on the full data set only omitting the final $T_{forecast}$ years as the test set. The remaining tests each generate a new test data set by iteratively removing more of the observations from the end of the data set. The number removed between each test can be controlled by changing the spacing argument. The kfold_cv function performs a k-fold version of leave_future_out_cv.

UniversalDiffEq.leave_future_out_cv — Method

leave_future_out_cv(model::UDE; kwargs ...)

Runs K fold leave future out cross validation and returns the mean squared forecasting error and a plot to visualize the model fits.

...

Arguments

model - the UDE model to test forecastlength = 5 - The number of data points to forecast forecastnumber = 10 - The number of trainin gnad testing data sets to create spacing = 2 - The number of data points to skip between the end of each new triaining data set BFGS=false - use BFGS algorithm to train the model stepsize = 0.05 - step size for first run of the ADAM algorithm maxiter = 500 - maximum iterations for first trial of ADAM stepsize2 = 0.01 - step size for second iteration of ADAM, only used of BFGS is false maxiter2 = 500 - step size for second iteration of ADAM, only used of BFGS is false ...

UniversalDiffEq.kfold_cv — Method

kfold_cv(model::UDE; kwargs ...)

Runs K fold leave future out cross validation and returns the mean squared forecasting error and a plot to visualize the model fits.

...

Arguments

model - a UDE model object k=10 - the number of testing and taining data sets to build leaveout=5 - the number of data points to leave out in each testingdata set BFGS = false - use the BFGS algorithm to train the model stepsize = 0.05 - step size for first run of the ADAM algorithm maxiter = 500 - maximum iterations for first trial of ADAM step_size2 = 0.01 - step size for second iteration of ADAM, only used of BFGS is false maxiter2 = 500 - step size for second iteration of ADAM, only used of BFGS is false ...