Model Constructors

UniversalDiffEq.jl provides a set of functions to construct NODEs and UDEs with varying levels of customization. The model constructors all require the data to be passed using a DataFrame object from the DataFrames.jl library. The data frame should be organized with a column for time named t, and the remaining columns should have the values of the state variables (X_i) at each point in time.

t	$X_1$	$X_2$

0.1	0.0	-1.1

0.2	0.01	-0.9

0.5	0.51	-1.05

Table: Example data set with two state variables

Currently, missing data are not supported, but irregular intervals between time points are allowed.

Each constructor function will require additional inputs to specify the model structure. For example, the CustomDerivatives function requires the user to supply the known functional forms through the derivs! argument. These arguments are described in detail in the subsection for each model type.

Finally, the constructor functions share a set of keyword arguments used to tune the model fitting procedure. These control the weights given to the process model, observation model, and regularization in the loss function and can be tuned to control the complexity of the estimated model and to accommodate varying levels of observational errors:

proc_weight=1.0 : The weight given to the model predictions in loss function

obs_weight=1.0 : The weight given to the state estimates in loss function

reg_weight=0.0 : The weight given to regularization in the loss function.

In addition to these weighting parameters, two keyword arguments, l = 0.25 and extrap_rho = 0.0, control how the model extrapolates beyond the observed data. The parameter l defines how far away the model will extrapolate before shifting to the default behavior and extrap_rho defines the default when extrapolating. When forecasting, the model will modify the trained process model $f(u_t;\theta)$ when extrapolating to a new function that combines the fitted model and the default behavior

\[\bar{f}(u_t|\theta,l,\rho )= \left\{ \begin{array}{ll} f(u_t;\theta) & min(\hat{u}) < u_t < max(\hat{u}) \\ e^{(\frac{u_t - min(\hat{u}_t)}{l})^2}f(u_t;\theta) + (1-e^{(\frac{u_t - min(\hat{u}_t)}{l})^2}) \rho &u_t < min(\hat{u}) \\ e^{(\frac{u_t - max(\hat{u}_t)}{l})^2}f(u_t;\theta) - (1-e^{(\frac{u_t - min(\hat{u}_t)}{l})^2}) \rho &u_t > max(\hat{u}) \\ \end{array} \right. \]

NODEs and NNDE

NODEs and NNDEs use neural networks to build fully nonparametric models in continuous and discrete time, respectively. NODEs use a neural network as the right-hand side of an ordinary differential equation

\[ \frac{dx}{dt} = NN(x;w,b),\]

and NNDEs use a neural network as the right-hand side of a difference equation

\[ x_{t+1} = x_t + NN(x_t).\]

The NODE and NNDE functions construct each model type.

UniversalDiffEq.NODE — Method

NODE(data;kwargs ... )

Constructs a nonparametric continuous time model for the data set data using a single layer neural network to represent the systems dynamics.

# kwargs

time_column_name: Name of column in data that corresponds to time. Default is "time".

hidden_units: Number of neurons in hidden layer. Default is 10.

seed: Fixed random seed for repeatable results. Default is 1.

proc_weight: Weight of process error omega_{proc}. Default is 1.0.

obs_weight: Weight of observation error omega_{obs}. Default is 1.0.

reg_weight: Weight of regularization error omega_{reg}. Default is 10^-6.

reg_type: Type of regularization, whether "L1" or "L2" regularization. Default is "L2".

l: Extrapolation parameter for forecasting. Default is 0.25.

extrap_rho: Extrapolation parameter for forecasting. Default is 0.0.

UniversalDiffEq.NNDE — Method

NNDE(data;kwargs ...)

Constructs a nonparametric discrete time model for the data set data using a single layer neural network to represent the systems dynamics.

# kwargs

time_column_name: Name of column in data that corresponds to time. Default is "time".

hidden_units: Number of neurons in hidden layer. Default is 10.

seed: Fixed random seed for repeatable results. Default is 1.

proc_weight: Weight of process error omega_{proc}. Default is 1.0.

obs_weight: Weight of observation error omega_{obs}. Default is 1.0.

reg_weight: Weight of regularization error omega_{reg}. Default is 10^-6.

reg_type: Type of regularization, whether "L1" or "L2" regularization. Default is "L2".

l: Extrapolation parameter for forecasting. Default is 0.25.

extrap_rho: Extrapolation parameter for forecasting. Default is 0.0.

Covariates can be added to the model by supplying a second data frame X. This data frame must have the same format as the primary data set, but the time points need not match. The NODE and NNDE functions will append the value of the covariates at each point in time to the neural network inputs

\[ \frac{dx}{dt} = NN(x,X(t);w,b) \\ x_{t+1} = x_t + NN(x_t, X(t)).\]

The values of the covariates between time points included in the data frame X are interpolated using a linear spline.

UniversalDiffEq.NODE — Method

NODE(data,X;kwargs ... )

When a dataframe X is supplied the model will run with covariates. the argument X should have a column for time t with the value for time in the remaining columns. The values in X will be interpolated with a linear spline for values of time not included in the data frame.

time_column_name: Name of column in data that corresponds to time. Default is "time".

variable_column_name: Name of column in data that corresponds to the variables. Default is "variable".

value_column_name: Name of column in data that corresponds to the covariates. Default is "value".

hidden_units: Number of neurons in hidden layer. Default is 10.

seed: Fixed random seed for repeatable results. Default is 1.

proc_weight: Weight of process error omega_{proc}. Default is 1.0.

obs_weight: Weight of observation error omega_{obs}. Default is 1.0.

reg_weight: Weight of regularization error omega_{reg}. Default is 10^-6.

reg_type: Type of regularization, whether "L1" or "L2" regularization. Default is "L2".

l: Extrapolation parameter for forecasting. Default is 0.25.

extrap_rho: Extrapolation parameter for forecasting. Default is 0.0.

UDEs

Continuous time model

The CustomDerivatives and CustomDifference functions can be used to build models that combine neural networks and known functional forms. These functions take user-defined models, construct a loss function, and provide access to the model fitting and testing functions provided by UniversalDiffEq.jl.

The CustomDerivatives function builds UDE models based on a user-defined function derivs!(du,u,p,t), which updates the vector du with the right-hand side of a differential equation evaluated at time t in state u given parameters p. The function also needs an initial guess at the model parameters, specified by a NamedTuple initial_parameters

UniversalDiffEq.CustomDerivatives — Method

CustomDerivatives(data,derivs!,initial_parameters;kwargs ... )

Constructs a UDE model for the data set data based on user defined derivatives derivs. An initial guess of model parameters are supplied with the initial_parameters argument.

...

Arguments

data: a DataFrame object with the time of observations in a column labeled t and the remaining columns the value of the state variables at each time point.

derivs: a Function of the form derivs!(du,u,p,t) where u is the value of the state variables, p are the model parameters, t is time, and du is updated with the value of the derivatives

init_parameters: A NamedTuple with the model parameters. Neural network parameters must be listed under the key NN.

kwargs

time_column_name: Name of column in data that corresponds to time. Default is "time".

proc_weight: Weight of process error omega_{proc}. Default is 1.0.

obs_weight: Weight of observation error omega_{obs}. Default is 1.0.

reg_weight: Weight of regularization error omega_{reg}. Default is 10^-6.

reg_type: Type of regularization, whether "L1" or "L2" regularization. Default is "L2".

l: Extrapolation parameter for forecasting. Default is 0.25.

extrap_rho: Extrapolation parameter for forecasting. Default is 0.0.

...

Example

The following block of code shows how to build a UDE model based on the Lotka-Volterra predator-prey model where the growth rate of the prey $r$, mortality rate of the predator $m$, and conversion efficiency $\theta$ are estimated and the predation rate is described by a neural network $NN$. The resulting ODE is defined by

\[\frac{dN}{dt} = rN - NN(N,P) \\ \frac{dP}{dt} = \theta NN(N,P) - mP.\]

To implement the model we start by defining the neural network object using the Lux.Chain function,

using Lux
# Build the neural network with Lux.Chain 
dims_in = 2
hidden_units = 10
nonlinearity = tanh
dims_out = 1
NN = Lux.Chain(Lux.Dense(dims_in,hidden_units,nonlinearity),Lux.Dense(hidden_units,dims_out))

# initialize the neural network states and parameters 
using Random
rng = Random.default_rng() 
NNparameters, states = Lux.setup(rng,NN) 

Now we can define the model derivatives using the usual Julia functions syntax. The derivs function first evaluates the neural network given the abundance of the predators and prey in the vector u. The neural network function NN requires three arguments: the current state, the network parameters, and the network states. In this example, the weights and biases are accessed through the parameters NamedTuple p with the key NN. The other model parameters are accessed with keys corresponding to their respective names.

function lotka_volterra_derivs!(du,u,p,t)
    C, states = NN(u,p.NN,states) 
    du[1] = p.r*u[1] - C[1]
    du[2] = p.theta*C[1] -p.m*u[2]
end

Finally, we can define the initial parameters as a NamedTuple and build the model using the CustomDerivatives function.

using UniversalDiffEq
initial_parameters = (NN = NNparameters, r = 1.0, m = 0.5, theta = 0.5)
model = CustomDerivatives(data,lotka_volterra_derivs!,initial_parameters)

Discrete time model

Discrete time models are constructed in a similar way to continuous time models. The user provides the right-hand side of a difference equation with the function step and initial parameters. The function step(u,t,p) takes three arguments: the value of the state variables u, time t, and model parameters p.

UniversalDiffEq.CustomDifference — Method

CustomDifference(data,step,initial_parameters;kwrags...)

Constructs a UDE model for the data set data based on user defined difference equation step. An initial guess of model parameters are supplied with the initial_parameters argument. ...

Arguments

data: a DataFrame object with the time of observations in a column labeled t and the remaining columns the value of the state variables at each time point.

step: a Function of the form step(u,t,p) where u is the value of the state variables, p are the model parameters.

init_parameters: A NamedTuple with the model parameters. Neural network parameters must be listed under the key NN.

kwargs

time_column_name: Name of column in data that corresponds to time. Default is "time".

proc_weight: Weight of process error omega_{proc}. Default is 1.0.

obs_weight: Weight of observation error omega_{obs}. Default is 1.0.

reg_weight: Weight of regularization error omega_{reg}. Default is 10^-6.

reg_type: Type of regularization, whether "L1" or "L2" regularization. Default is "L2".

l: Extrapolation parameter for forecasting. Default is 0.25.

extrap_rho: Extrapolation parameter for forecasting. Default is 0.0.

...

Adding covariates

Covariates can also be added to UDE models by passing a data frame X and adding covariates as an argument to the derivs! function which has the new form derivs!(du,u,X,p,t), where the third argument X is a vector of covariates.

UniversalDiffEq.CustomDerivatives — Method

CustomDerivatives(data::DataFrame,X,derivs!::Function,initial_parameters;kwargs ... )

When a dataframe X is supplied the model will run with covariates. the argument X should have a column for time t with the value for time in the remaining columns. The values in X will be interpolated with a linear spline for value of time not included in the data frame.

When X is provided the derivs function must have the form derivs!(du,u,x,p,t) where x is a vector with the value of the covariates at time t.

kwargs

time_column_name: Name of column in data that corresponds to time. Default is "time".

variable_column_name: Name of column in data that corresponds to the variables. Default is "variable".

value_column_name: Name of column in data that corresponds to the covariates. Default is "value".

proc_weight: Weight of process error omega_{proc}. Default is 1.0.

obs_weight: Weight of observation error omega_{obs}. Default is 1.0.

reg_weight: Weight of regularization error omega_{reg}. Default is 10^-6.

reg_type: Type of regularization, whether "L1" or "L2" regularization. Default is "L2".

l: Extrapolation parameter for forecasting. Default is 0.25.

extrap_rho: Extrapolation parameter for forecasting. Default is 0.0.

Covariates can also be added to a discrete time framework in the same way. The step function should have four arguments step(u,X,t,p).

UniversalDiffEq.CustomDifference — Method

CustomDifference(data::DataFrame,X,step,initial_parameters;kwargs ... )

When a dataframe X is supplied the model will run with covariates. the argument X should have a column for time t with the value for time in the remaining columns. The values in X will be interpolated with a linear spline for value of time not included in the data frame.

When X is provided the step function must have the form step(u,x,t,p) where x is a vector with the value of the covariates at time t.

# kwargs

time_column_name: Name of column in data that corresponds to time. Default is "time".

variable_column_name: Name of column in data that corresponds to the variables. Default is "variable".

value_column_name: Name of column in data that corresponds to the covariates. Default is "value".

proc_weight: Weight of process error omega_{proc}. Default is 1.0.

obs_weight: Weight of observation error omega_{obs}. Default is 1.0.

reg_weight: Weight of regularization error omega_{reg}. Default is 10^-6.

reg_type: Type of regularization, whether "L1" or "L2" regularization. Default is "L2".

l: Extrapolation parameter for forecasting. Default is 0.25.

extrap_rho: Extrapolation parameter for forecasting. Default is 0.0.

Example

To show how adding covariates can work, the following example extends the Lotka-Volterra equations defined above to incorporate a covariate X that influences the abundance of predators and prey. We can model this as a linear effect with coefficients $\beta_N$ and $\beta_P$

\[\frac{dN}{dt} = rN - NN(N,P) + \beta_N N \\ \frac{dP}{dt} = \theta NN(N,P) - mP + \beta_P P.\]

# set up neural network 
using Lux
dims_in = 2
hidden_units = 10
nonlinearity = tanh
dims_out = 1
NN = Lux.Chain(Lux.Dense(dims_in,hidden_units,nonlinearity),Lux.Dense(hidden_units,dims_out))

# initialize parameters 
using Random
rng = Random.default_rng() 
NNparameters, NNstates = Lux.setup(rng,NN) 

function derivs!(du,u,X,p,t)
    C, states = NN(u,p.NN, NNstates) # NNstates are
    du[1] = p.r*u[1] - C[1] + p.beta[1] * X[1]
    du[2] = p.theta*C[1] -p.m*u[2] + p.beta[2] * X[1]
end

init_parameters = (NN = NNparameters, r = 1.0, m = 0.5, theta = 0.5, beta = [0,0])

model = CustomDerivatives(training_data,X,derivs!;init_parameters;proc_weight=2.0,obs_weight=0.5,reg_weight=10^-4)
nothing

Covariates with different sampling frequencies

If you wish to build a model with covariates that are measured at different points in time, then you can provide a list of data frames as the covariates argument. Each data frame in the list should have time in the first column and the value of one of the covariates in the second column. The model will interpolate each time series with a linear spline. The value of the covariates can be accessed in the derivs and step functions for custom models by indexing into the covariates argument X. The values will be listed in the same order as the data frames are provided in the covariates argument.

Adding prior information to custom models

UniversalDiffEq.CustomDerivatives — Method

CustomDerivatives(data::DataFrame,derivs!::Function,initial_parameters,priors::Function;kwargs ... )

When a function's priors is supplied its value will be added to the loss function as a penalty term for user specified parameters. It should take the a single NamedTuple p as an argument penalties for each paramter should be calculated by accessing p with the period operator.

The prior function can be used to nudge the fitted model toward prior expectations for a parameter value. For example, the following function increases the loss when a parameter p.r has a value other than 1.5, nad a second parameter p.beta is greater than zeros.

function priors(p)
    l = 0.01*(p.r - 1.5)^2
    l += 0.01*(p.beta)^2
    return l
end 

kwargs

time_column_name: Name of column in data that corresponds to time. Default is "time".

proc_weight: Weight of process error omega_{proc}. Default is 1.0.

obs_weight: Weight of observation error omega_{obs}. Default is 1.0.

reg_weight: Weight of regularization error omega_{reg}. Default is 10^-6.

reg_type: Type of regularization, whether "L1" or "L2" regularization. Default is "L2".

l: Extrapolation parameter for forecasting. Default is 0.25.

extrap_rho: Extrapolation parameter for forecasting. Default is 0.0.

UniversalDiffEq.CustomDifference — Method

CustomDifference(data::DataFrame,step,initial_parameters,priors::Function;kwargs ... )

When a function's priors is supplied its value will be added to the loss function as a penalty term for user specified paramters. It should take the a single NamedTuple p as an argument penalties for each parameter should be calcualted by accessing p with the period operator.

function priors(p)
    l = 0.01*(p.r - 1.5)^2
    l += 0.01*(p.beta)^2
    return l
end 

kwargs

time_column_name: Name of column in data that corresponds to time. Default is "time".

proc_weight: Weight of process error omega_{proc}. Default is 1.0.

obs_weight: Weight of observation error omega_{obs}. Default is 1.0.

reg_weight: Weight of regularization error omega_{reg}. Default is 10^-6.

reg_type: Type of regularization, whether "L1" or "L2" regularization. Default is "L2".

l: Extrapolation parameter for forecasting. Default is 0.25.

extrap_rho: Extrapolation parameter for forecasting. Default is 0.0.