Fitting a model to multiple time series

UniversalDiffEq.jl provides a set of functions to fit models to multiple time series. The functions for this mirror the functions for fitting NODEs and UDEs to single time series with the prefix Multi. For example, to build a NODE model for multiple time series you would use the MultiNODE function. The functions for model fitting, testing, and visualization have the same names. The other important difference is the data format: a column with a unique index for each time series must be included.

Dataframe

t	series	x1	x2

1	1	$x_{1,1,t}$	$x_{1,2,t}$

2	1	$x_{1,1,t}$	$x_{1,2,t}$

3	1	$x_{1,1,t}$	$x_{1,2,t}$

1	2	$x_{2,1,t}$	$x_{2,2,t}$

2	2	$x_{2,1,t}$	$x_{2,2,t}$

3	2	$x_{2,1,t}$	$x_{2,2,t}$

Covariates can be added to the models as well. The covariates data frame must have the same structure.

UniversalDiffEq.MultiNODE — Method

MultiNODE(data;kwargs...)

Builds a NODE model to fit to the data with multiple time series. data is a DataFrame object with time arguments placed in a column labeled t and a second column with a unique index for each time series. The remaining columns have observations of the state variables at each point in time and for each time series.

UniversalDiffEq.MultiNODE — Method

MultiNODE(data,X;kwargs...)

When a dataframe X is supplied, the model will run with covariates. The argument X should have a column for time, a column for series index, a column for covariate names, and a column for the covariate values at each time step. The values in X will be interpolated with a linear spline for values of time not included in the dataframe.