UDE model construction

Some users may wish to access elements of a fitted model directly to create custom model visualizations, performance tests, or other applications not forseen by thr developers. To this end, we provide documentation of classes (Julia mutable structs) used by UniversalDiffEq.jl to build the NODE and UDE objects. The package is built around the UDE class which stores the data used to fit a model and instances of six submodel classes used to define the full model.

UniversalDiffEq uses a state space modeling framework to define and fit NODE and UDE models. State space models are a class of time series models that describe a time series data with a process model that describes the dynaics of a sequence of unobserved state variables $u_t$ a second observaiton model defines the relationship between the state variables $u_t$ and the observations $x_t$. The process model $f$ predicts value of the state variables one step ahead

\[\hat{u}_{t+\Delta t } = f(u_t; t, \Delta t, \theta_{proc})\]

where $\Delta t$ is the time span between observations, and $\theta_{proc}$ is the model paramters. The observaiton model maps from the state variables $u_t$ to the observations

\[x_{t} = h(u_t; t, \Delta t, \theta_{obs})\]

where $\theta_{obs}$ are the observaiton model parameters. In addition to these primary functions both the observaiton models and process models have loss funtion to measure the accuracy of thier predictions. These can be thought of like the likelihood models used in generalized linear models. for example, we can measure the perforance of the process model with a normal likelihood

\[L(\hat{u}_t,u_t) = \frac{1}{\sigma \sqrt{2\pi}} e^{-\frac{1}{2}(\frac{\hat{u}_t-u_t}{\sigma})^2}\]

where $\sigma$ is the variance of the prediciton errors. Although in principal any likelihood can be used, we use the mean squared error in our base model specification.

The UDE models also include submodels to regualrize the process and observaiton models. The regualrization models are functions of the model parameters that add to the loss funtion. The regularizaiton models are ineffect priors on the model parameters. Regularization in expecially important for nerual network models. For example, out default model constructors apply L2 regualrizaiton to neuarl network paramters in the process model

\[R(\theta_{proc}) = \omega ||\theta_{proc}||_{L2}\]

where $\omega$ is the weight given to regualrization in the over all loss function.

These six model components are all combined into one loss functions used to fit the UDE models

\[L(u,\theta_{proc},\theta_{obs};x) = \sum_{t =1}^{T} L_{obs}(x_t,h(u_t,\theta_{obs});\sigma_{obs}) + \sum_{t=2}^{T} L_{proc}(u_t,f(u_{t-1},\theta_{proc});\sigma_{proc}) + R_{obs}(\theta_{obs}) + R_{proc}(\theta_{proc}).\]

where the $\sigma_i$ are paramters for the loss functions and the $\theta_i$ are paramters for the prediction functions.

The UDE object combines the observation and process models and their rpective loss and regualrization models into one larger model object along with the data used to fit the model.

UniversalDiffEq.UDE — Type

UDE

Basic data structure used to the model structure, parameters and data for UDE and NODE models. ...

Elements

times: a vector of times for each observation

data: a matrix of observaitons at each time point

X: a DataFrame with any covariates used by the model

data_frame: a DataFrame with colums for the time of each observation and values of the state variables

parameters: a ComponentArray that stores model parameters

loss_function: the loss function used to fit the model

process_model: a Julia mutable struct used to define model predictions

process_loss: a Julia mutable struct used to measure the peroance of model predictions

observation_model: a Julia mutable struct used to predict observaitons given state variable estiamtes

observaiton_loss: a Julia mutable struct used to measure the performance of the observaiton model

process_regularization: a Julia mutable struct used to store data needed for process model regularization

observation_regularization: a Julia mutable struct used to store data needed for observation model regularization

constructor: A function that initializes a UDE model with identical structure.

...

UniversalDiffEq.ProcessModel — Type

ProcessModel

A Julia mutable struct that stores the functions and parameters for the process model. ...

Elements

parameters: ComponentArray

predict: Function the predict one time step ahead

forecast: Function, a modified version of rpedict to imporve performace when extrapolating

covariates: Function that returns the value of the covariates at each point in time.

...

UniversalDiffEq.LossFunction — Type

LossFunction

A Julia mutable struct that stores the loss function and parameters. ...

Elements

parameters: ComponentArray

loss: Function

...

UniversalDiffEq.Regularization — Type

Regularization

A Julia mutable struct that stores the loss function and parameters. ...

Elements

reg_parameters: ComponentArray

loss: Function

...