Model analysis

UniversalDiffEq.jl provides several functions to analyze the characteristics of the fitted models. The most basic of these is the get_right_hand_side function. This function takes a UDE model as an argument and returns the right-hand side of the fitted differential or difference equation. This function can then be treated like any dynamic model and analyzed for equilibria, stability, tipping points, and other phenomena of interest.

UniversalDiffEq.get_right_hand_side — Method

get_right_hand_side(UDE::UDE)

Returns the right hand side of the differential equation (or difference equation) used to build the process model.

The fuction will take the state vector u and time t if the model does not include covariates. If covaraites are included the arguments are the state vector u , covariates vector x, and time t

The library also has functions to evaluate model predictions. The forecast function will run a simulation of the model starting at the initial point u0 and returning the value of the state variables at each point in the times vector.

UniversalDiffEq.forecast — Method

forecast(UDE::UDE, u0::AbstractVector{}, times::AbstractVector{})

predicitons from the trained model UDE starting at u0 saving values at times. Assumes u0 is the value at time times[1]

UniversalDiffEq.print_parameter_estimates — Method

print_parameter_estimates(UDE::UDE)

prints the value of the known dynamcis parameters.

UniversalDiffEq.plot_forecast — Method

plot_forecast(UDE::UDE, T::Int)

Plots the models forecast up to T time steps into the future from the last observation.

UniversalDiffEq.get_NN_parameters — Method

get_NN_parameters(UDE::UDE)

Returns value weights and biases of the neural network

UniversalDiffEq.get_parameters — Method

get_parameters(UDE::UDE)

Returns model parameters.