Easy NODE and UDE

EasyNODE and EasyUDE provide a quick, simple alternative to the other model constructors featured by UniversalDiffEq.jl. They each return pre-trained models, in which neural networks are kept to 1 hidden layer.

EasyNODE constructors:

UniversalDiffEq.EasyNODE — Method

EasyNODE(data,X;kwargs ... )

When a dataframe X is supplied the model will run with covariates. the argument X should have a column for time t with the value for time in the remaining columns. The values in X will be interpolated with a linear spline for values of time not included in the data frame.

UniversalDiffEq.EasyNODE — Method

EasyNODE(data;kwargs ... )

Constructs a pretrained continuous time model for the data set data using a single layer neural network to represent the systems dynamics.

EasyUDE constructors:

UniversalDiffEq.EasyUDE — Method

EasyUDE(data,derivs!,initial_parameters;kwargs ... )

Constructs a pretrained UDE model for the data set data based on user defined derivatives derivs. An initial guess of model parameters are supplied with the initial_parameters argument.

UniversalDiffEq.EasyUDE — Method

EasyUDE(data::DataFrame,X,derivs!::Function,initial_parameters;kwargs ... )

When a dataframe X is supplied the model will run with covariates. the argument X should have a column for time t with the value for time in the remaining columns. The values in X will be interpolated with a linear spline for value of time not included in the data frame. When X is provided the derivs function must have the form derivs!(du,u,x,p,t) where x is a vector with the value of the covariates at time t.