Model Constructors

UniversalDiffEq provides a set of functions to construct NODE and UDE with varying levels of customization.

NODES and NNDE

The simplest models to implement are the fully nonparametrics neural ordinary differential equation (NODE) and neural newtork difference equation (NNDE). These functions use a neural network to represent the right hand side of a system of differential equations and differnce equation respectively

\[ \frac{dx}{dt} = NN(x;w,b) \\ x_{t+1} = NN(x_t;w,b)\]

The NODE function builds a NODE model for a user supplied data set and NNDE build

UniversalDiffEq.NODE — Method

NODE(data;kwargs ... )

Constructs a nonparametric continuous time model for the data set data using a single layer neural network to represent the systems dynamics.

Model equations

\[ dx/dt = NN(x;w,b)\]

...

Arguments

data: a DataFrame object with the time of observations in a column labeled t and the remaining columns the value of the state variables at each time point.

Key word arguments

proc_weight=1.0 : Weight given to the model predictiosn in loss funciton

obs_weight=1.0 : Weight given to the state estiamtes in loss function

reg_weight=10^-6 : Weight given to regularization in the loss function

extrap_rho=0.0 : Asymthotic value of derivitives when extrapolating (negative when extrapolating higher than past observaitons, postive when extrapolating lower)

l=0.25 : rate at which extrapolations converge on asymthotic behavior

...

Covariates can be added to the model by supplying a second data frame X with thier values at differnt points in time. The value of the covaritates are appended to the state vector $x$ to create the input for the neural network

\[ \frac{dx}{dt} = NN(x,X(t);w,b) \\ x_{t+1} = NN(x_t,X_t;w,b)\]

The NODE models use a linear interpolation to approximate the value fo the covarites between observations. odel with covariates have not been developed for the discrete time case yet.

UniversalDiffEq.NODE — Method

NODE(data,X;kwargs ... )

When a dataframe X is supplied the model will run with covariates. the argumetn X should have a column for time t with the vlaue fo time in the remaining columns. The values in X will be interpolated with a linear spline for value of time not included int he data frame.

When X is provided the derivs function must have the form derivs!(du,u,x,p,t) where x is a vector with the value of the coarates at time t.

UDEs

The CustomDerivatives and CustomDifference function can be used to build models that combine nerual networks and known functional forms. These function take user defined models and consturct a loss function and provide access to the model fitting and testing functions provided by UniversalDiffEq.jl

UniversalDiffEq.CustomDerivatives — Method

CustomDerivatives(data,derivs!,initial_parameters;kwargs ... )

Constructs a UDE model for the data set data based on user defined derivitivs derivs. An initial guess of model parameters are supplied with the initia_parameters argument.

...

Arguments

data: a DataFrame object with the time of observations in a column labeled t and the remaining columns the value of the state variables at each time point.

derivs: a Function of the form derivs!(du,u,p,t) where u is the value of the state variables, p are the model parameters, t is time, and du is updated with the value of the derivitives

init_parameters: A NamedTuple with the model parameters. Neural network parameters must be listed under the key NN.

Key word arguments

proc_weight=1.0 : Weight given to the model predictiosn in loss funciton

obs_weight=1.0 : Weight given to the state estiamtes in loss function

reg_weight=10^-6 : Weight given to regularization in the loss function

extrap_rho=0.0 : Asymthotic value of derivitives when extrapolating (negative when extrapolating higher than past observaitons, postive when extrapolating lower)

l=0.25 : rate at which extrapolations converge on asymthotic behavior

...

UniversalDiffEq.CustomDiffernce — Method

CustomDiffernce(data,step,initial_parameters;kwrags...)

Constructs a UDE model for the data set data based on user defined difference equation step. An initial guess of model parameters are supplied with the initia_parameters argument.

data: a DataFrame object with the time of observations in a column labeled t and the remaining columns the value of the state variables at each time point.

step: a Function of the form step(u,t,p) where u is the value of the state variables, p are the model parameters.

init_parameters: A NamedTuple with the model parameters. Neural network parameters must be listed under the key NN.

...

Key word arguments

proc_weight=1.0 : Weight given to the model predictiosn in loss funciton

obs_weight=1.0 : Weight given to the state estiamtes in loss function

reg_weight=10^-6 : Weight given to regularization in the loss function

extrap_rho=0.0 : Asymthotic value of derivitives when extrapolating (negative when extrapolating higher than past observaitons, postive when extrapolating lower)

l=0.25 : rate at which extrapolations converge on asymthotic behavior

...

Adding Covariates

In this context the

UniversalDiffEq.CustomDerivatives — Method

CustomDerivatives(data,X,derivs!,initial_parameters;kwargs ... )

When a dataframe X is supplied the model will run with covariates. the argumetn X should have a column for time t with the vlaue fo time in the remaining columns. The values in X will be interpolated with a linear spline for value of time not included int he data frame.

When X is provided the derivs function must have the form derivs!(du,u,x,p,t) where x is a vector with the value of the coarates at time t.

UniversalDiffEq.CustomDiffernce — Method

CustomDiffernce(data,X,step,initial_parameters;kwargs ... )

When a dataframe X is supplied the model will run with covariates. the argumetn X should have a column for time t with the vlaue fo time in the remaining columns. The values in X will be interpolated with a linear spline for value of time not included int he data frame.

When X is provided the step function must have the form step(u,x,t,p) where x is a vector with the value of the coarates at time t.

Other functions

UniversalDiffEq.NNDE — Method

NNDE(data;kwargs ...)

Constructs a nonparametric discrete time model for the data set data using a single layer neural network to reporesent the systems dynamics.

Model equations

\[x_{t+1} = NN(x_t;w,b)\]

...

Arguments

data: a DataFrame object with the time of observations in a column labeled t and the remaining columns the value of the state variables at each time point.

Key word arguments

proc_weight=1.0 : Weight given to the model predictiosn in loss funciton

obs_weight=1.0 : Weight given to the state estiamtes in loss function

reg_weight=10^-6 : Weight given to regularization in the loss function

extrap_rho=0.0 : Asymthotic value of derivitives when extrapolating (negative when extrapolating higher than past observaitons, postive when extrapolating lower)

l=0.25 : rate at which extrapolations converge on asymthotic behavior

...

UniversalDiffEq.DiscreteUDE — Method

DiscreteUDE(data,step,init_parameters;kwargs ...)

Constructs an additive UDE model with user supplied difference equations step and a single layer neural network. When init_parameters are provided for the use supplied function their values will be estiated in the training process.

Model equaitons

\[x_{t+1} = f(x_t; heta) + NN(x_t;w,b)\]

...

Key word arguments

proc_weight=1.0 : Weight given to the model predictiosn in loss funciton

obs_weight=1.0 : Weight given to the state estiamtes in loss function

reg_weight=10^-6 : Weight given to regularization in the loss function

extrap_rho=0.0 : Asymthotic value of derivitives when extrapolating (negative when extrapolating higher than past observaitons, postive when extrapolating lower)

l=0.25 : rate at which extrapolations converge on asymthotic behavior

...

UniversalDiffEq.UDE — Method

UDE(data,derivs,init_parameters;kwargs...)

Constructs an additive continuous time UDE model with user supplied derivitives step and a single layer neural network. When init_parameters are provided for the user supplied function their values will be estiated during model training.

Model equaitons

\[dx/dt = = f(x; heta) + NN(x;w,b)\]

...

Key word arguments

proc_weight=1.0 : Weight given to the model predictiosn in loss funciton

obs_weight=1.0 : Weight given to the state estiamtes in loss function

reg_weight=10^-6 : Weight given to regularization in the loss function

extrap_rho=0.0 : Asymthotic value of derivitives when extrapolating (negative when extrapolating higher than past observaitons, postive when extrapolating lower)

l=0.25 : rate at which extrapolations converge on asymthotic behavior

...